requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
dataclasses-json==0.6.3
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docx==0.2.4
executing==2.0.1
fastapi==0.109.2
filelock==3.13.1
Flask==2.2.5
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kubernetes==29.0.0
langchain==0.1.8
langchain-community==0.0.29
langchain-core==0.1.34
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
langsmith==0.1.5
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.10.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
posthog==3.4.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
setuptools==69.0.3
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.36.3
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.38.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.27.1
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
yarl==1.9.4
youtube-comment-downloader==0.1.70
zipp==3.17.0
